Move gamepads handling to the separate file
