+++ apps/openmw/mwinput/inputmanagerimp.cpp
@@ -31,6 +31,7 @@
 #include "../mwmechanics/actorutil.hpp"
 
 #include "actionmanager.hpp"
+#include "controllermanager.hpp"
 #include "mousemanager.hpp"
 #include "sdlmappings.hpp"
 #include "sensormanager.hpp"
@@ -46,31 +47,23 @@ namespace MWInput
             const std::string& controllerBindingsFile, bool grab)
         : mWindow(window)
         , mWindowVisible(true)
-        , mJoystickLastUsed(false)
         , mInputManager(nullptr)
         , mVideoWrapper(nullptr)
         , mUserFile(userFile)
         , mDragDrop(false)
         , mGrabCursor (Settings::Manager::getBool("grab cursor", "Input"))
         , mControlsDisabled(false)
-        , mJoystickEnabled (Settings::Manager::getBool("enable controller", "Input"))
         , mPreviewPOVDelay(0.f)
         , mTimeIdle(0.f)
         , mGuiCursorEnabled(true)
-        , mGamepadGuiCursorEnabled(true)
         , mDetectingKeyboard(false)
         , mOverencumberedMessageDelay(0.f)
-        , mGamepadZoom(0)
-        , mSneakToggleShortcutTimer(0.f)
-        , mSneakGamepadShortcut(false)
         , mAttemptJump(false)
-        , mGamepadCursorSpeed(Settings::Manager::getFloat("gamepad cursor speed", "Input"))
         , mFakeDeviceID(1)
     {
         mInputManager = new SDLUtil::InputWrapper(window, viewer, grab);
         mInputManager->setKeyboardEventCallback (this);
         mInputManager->setWindowEventCallback(this);
-        mInputManager->setControllerEventCallback(this);
 
         mVideoWrapper = new SDLUtil::VideoWrapper(window, viewer);
         mVideoWrapper->setGammaContrast(Settings::Manager::getFloat("gamma", "Video"),
@@ -95,42 +88,16 @@ namespace MWInput
         mControlSwitch["playerviewswitch"]    = true;
         mControlSwitch["vanitymode"]          = true;
 
-        /* Joystick Init */
-
-        // Load controller mappings
-        if(!controllerBindingsFile.empty())
-        {
-            SDL_GameControllerAddMappingsFromFile(controllerBindingsFile.c_str());
-        }
-        if(!userControllerBindingsFile.empty())
-        {
-            SDL_GameControllerAddMappingsFromFile(userControllerBindingsFile.c_str());
-        }
-
-        // Open all presently connected sticks
-        int numSticks = SDL_NumJoysticks();
-        for(int i = 0; i < numSticks; i++)
-        {
-            if(SDL_IsGameController(i))
-            {
-                SDL_ControllerDeviceEvent evt;
-                evt.which = i;
-                controllerAdded(mFakeDeviceID, evt);
-                Log(Debug::Info) << "Detected game controller: " << SDL_GameControllerNameForIndex(i);
-            }
-            else
-            {
-                Log(Debug::Info) << "Detected unusable controller: " << SDL_JoystickNameForIndex(i);
-            }
-        }
-
-        mSensorManager = new SensorManager();
-        mInputManager->setSensorEventCallback (mSensorManager);
+        mActionManager = new ActionManager(mInputBinder, screenCaptureOperation, viewer, screenCaptureHandler);
 
         mMouseManager = new MouseManager(mInputBinder, mInputManager, window);
         mInputManager->setMouseEventCallback (mMouseManager);
 
-        mActionManager = new ActionManager(mInputBinder, screenCaptureOperation, viewer, screenCaptureHandler);
+        mControllerManager = new ControllerManager(mInputBinder, mInputManager, mActionManager, mMouseManager, userControllerBindingsFile, controllerBindingsFile);
+        mInputManager->setControllerEventCallback(mControllerManager);
+
+        mSensorManager = new SensorManager();
+        mInputManager->setSensorEventCallback (mSensorManager);
     }
 
     void InputManager::clear()
@@ -140,6 +107,7 @@ namespace MWInput
             it->second = true;
 
         mActionManager->clear();
+        mControllerManager->clear();
         mSensorManager->clear();
         mMouseManager->clear();
     }
@@ -149,6 +117,7 @@ namespace MWInput
         mInputBinder->save (mUserFile);
 
         delete mActionManager;
+        delete mControllerManager;
         delete mMouseManager;
         delete mSensorManager;
 
@@ -176,93 +145,6 @@ namespace MWInput
             int pc = playerChannels[i];
             mInputBinder->getChannel(pc)->setEnabled(enabled);
         }
-    }
-
-    bool InputManager::gamepadToGuiControl(const SDL_ControllerButtonEvent &arg)
-    {
-        // Presumption of GUI mode will be removed in the future.
-        // MyGUI KeyCodes *may* change.
-
-        MyGUI::KeyCode key = MyGUI::KeyCode::None;
-        switch (arg.button)
-        {
-            case SDL_CONTROLLER_BUTTON_DPAD_UP:
-                key = MyGUI::KeyCode::ArrowUp;
-                break;
-            case SDL_CONTROLLER_BUTTON_DPAD_RIGHT:
-                key = MyGUI::KeyCode::ArrowRight;
-                break;
-            case SDL_CONTROLLER_BUTTON_DPAD_DOWN:
-                key = MyGUI::KeyCode::ArrowDown;
-                break;
-            case SDL_CONTROLLER_BUTTON_DPAD_LEFT:
-                key = MyGUI::KeyCode::ArrowLeft;
-                break;
-            case SDL_CONTROLLER_BUTTON_A:
-                // If we are using the joystick as a GUI mouse, A must be handled via mouse.
-                if (mGamepadGuiCursorEnabled)
-                    return false;
-                key = MyGUI::KeyCode::Space;
-                break;
-            case SDL_CONTROLLER_BUTTON_B:
-                if (MyGUI::InputManager::getInstance().isModalAny())
-                    MWBase::Environment::get().getWindowManager()->exitCurrentModal();
-                else
-                    MWBase::Environment::get().getWindowManager()->exitCurrentGuiMode();
-                return true;
-            case SDL_CONTROLLER_BUTTON_X:
-                key = MyGUI::KeyCode::Semicolon;
-                break;
-            case SDL_CONTROLLER_BUTTON_Y:
-                key = MyGUI::KeyCode::Apostrophe;
-                break;
-            case SDL_CONTROLLER_BUTTON_LEFTSHOULDER:
-                key = MyGUI::KeyCode::Period;
-                break;
-            case SDL_CONTROLLER_BUTTON_RIGHTSHOULDER:
-                key = MyGUI::KeyCode::Slash;
-                break;
-            case SDL_CONTROLLER_BUTTON_LEFTSTICK:
-                mGamepadGuiCursorEnabled = !mGamepadGuiCursorEnabled;
-                MWBase::Environment::get().getWindowManager()->setCursorActive(mGamepadGuiCursorEnabled);
-                return true;
-            default:
-                return false;
-        }
-
-        // Some keys will work even when Text Input windows/modals are in focus.
-        if (SDL_IsTextInputActive())
-            return false;
-
-        MWBase::Environment::get().getWindowManager()->injectKeyPress(key, 0, false);
-        return true;
-    }
-
-    bool InputManager::gamepadToGuiControl(const SDL_ControllerAxisEvent &arg)
-    {
-        switch (arg.axis)
-        {
-            case SDL_CONTROLLER_AXIS_TRIGGERRIGHT:
-                if (arg.value == 32767) // Treat like a button.
-                    MWBase::Environment::get().getWindowManager()->injectKeyPress(MyGUI::KeyCode::Minus, 0, false);
-                break;
-            case SDL_CONTROLLER_AXIS_TRIGGERLEFT:
-                if (arg.value == 32767) // Treat like a button.
-                    MWBase::Environment::get().getWindowManager()->injectKeyPress(MyGUI::KeyCode::Equals, 0, false);
-                break;
-            case SDL_CONTROLLER_AXIS_LEFTX:
-            case SDL_CONTROLLER_AXIS_LEFTY:
-            case SDL_CONTROLLER_AXIS_RIGHTX:
-            case SDL_CONTROLLER_AXIS_RIGHTY:
-                // If we are using the joystick as a GUI mouse, process mouse movement elsewhere.
-                if (mGamepadGuiCursorEnabled)
-                    return false;
-                break;
-            default:
-                return false;
-        }
-
-        return true;
     }
 
     void InputManager::channelChanged(ICS::Channel* channel, float currentValue, float previousValue)
@@ -297,12 +179,13 @@ namespace MWInput
 
         if (mControlSwitch["playercontrols"])
         {
+            bool joystickUsed = mControllerManager->joystickLastUsed();
             if (action == A_Use)
             {
-                if(mJoystickLastUsed && currentValue == 1.0 && actionIsActive(A_ToggleWeapon))
+                if(joystickUsed && currentValue == 1.0 && actionIsActive(A_ToggleWeapon))
                     action = A_CycleWeaponRight;
 
-                else if (mJoystickLastUsed && currentValue == 1.0 && actionIsActive(A_ToggleSpell))
+                else if (joystickUsed && currentValue == 1.0 && actionIsActive(A_ToggleSpell))
                     action = A_CycleSpellRight;
 
                 else
@@ -314,10 +197,10 @@ namespace MWInput
             }
             else if (action == A_Jump)
             {
-                if(mJoystickLastUsed && currentValue == 1.0 && actionIsActive(A_ToggleWeapon))
+                if(joystickUsed && currentValue == 1.0 && actionIsActive(A_ToggleWeapon))
                     action = A_CycleWeaponLeft;
 
-                else if (mJoystickLastUsed && currentValue == 1.0 && actionIsActive(A_ToggleSpell))
+                else if (joystickUsed && currentValue == 1.0 && actionIsActive(A_ToggleSpell))
                     action = A_CycleSpellLeft;
 
                 else
@@ -371,28 +254,7 @@ namespace MWInput
 
         updateCursorMode();
 
-        if (mGuiCursorEnabled && !(mJoystickLastUsed && !mGamepadGuiCursorEnabled))
-        {
-            float xAxis = mInputBinder->getChannel(A_MoveLeftRight)->getValue()*2.0f-1.0f;
-            float yAxis = mInputBinder->getChannel(A_MoveForwardBackward)->getValue()*2.0f-1.0f;
-            float zAxis = mInputBinder->getChannel(A_LookUpDown)->getValue()*2.0f-1.0f;
-
-            xAxis *= (1.5f - mInputBinder->getChannel(A_Use)->getValue());
-            yAxis *= (1.5f - mInputBinder->getChannel(A_Use)->getValue());
-
-            // We keep track of our own mouse position, so that moving the mouse while in
-            // game mode does not move the position of the GUI cursor
-            float xMove = xAxis * dt * 1500.0f * mInvUiScalingFactor * mGamepadCursorSpeed;
-            float yMove = yAxis * dt * 1500.0f * mInvUiScalingFactor * mGamepadCursorSpeed;
-            if (xMove != 0|| yMove != 0 || zAxis != 0)
-            {
-                int mouseWheelMove = static_cast<int>(-zAxis * dt * 1500.0f);
-
-                mMouseManager->injectMouseMove(xMove, yMove, mouseWheelMove);
-                mMouseManager->warpMouse();
-                MWBase::Environment::get().getWindowManager()->setCursorActive(true);
-            }
-        }
+        mControllerManager->update(dt, disableControls, mPreviewPOVDelay == 1.f);
 
         if (mMouseManager->update(dt, disableControls))
             resetIdleTime();
@@ -414,28 +276,10 @@ namespace MWInput
                 bool isRunning = false;
                 bool alwaysRunAllowed = false;
 
-                // joystick movement
+                // keyboard movement
                 float xAxis = mInputBinder->getChannel(A_MoveLeftRight)->getValue();
                 float yAxis = mInputBinder->getChannel(A_MoveForwardBackward)->getValue();
-                if (xAxis != .5)
-                {
-                    triedToMove = true;
-                    player.setLeftRight((xAxis - 0.5f) * 2);
-                }
-
-                if (yAxis != .5)
-                {
-                    triedToMove = true;
-                    player.setAutoMove (false);
-                    player.setForwardBackward((yAxis - 0.5f) * 2 * -1);
-                }
-
-                if (triedToMove)
-                    mJoystickLastUsed = true;
-
-                // keyboard movement
                 isRunning = xAxis > .75 || xAxis < .25 || yAxis > .75 || yAxis < .25;
-                if(triedToMove) resetIdleTime();
 
                 if (actionIsActive(A_MoveLeft) != actionIsActive(A_MoveRight))
                 {
@@ -462,34 +306,7 @@ namespace MWInput
                 static const bool isToggleSneak = Settings::Manager::getBool("toggle sneak", "Input");
                 if (!isToggleSneak)
                 {
-                    if(mJoystickLastUsed)
-                    {
-                        if(actionIsActive(A_Sneak))
-                        {
-                            if(mSneakToggleShortcutTimer) // New Sneak Button Press
-                            {
-                                if(mSneakToggleShortcutTimer <= 0.3f)
-                                {
-                                    mSneakGamepadShortcut = true;
-                                    mActionManager->toggleSneaking();
-                                }
-                                else
-                                    mSneakGamepadShortcut = false;
-                            }
-
-                            if(!mActionManager->isSneaking())
-                                mActionManager->toggleSneaking();
-                            mSneakToggleShortcutTimer = 0.f;
-                        }
-                        else
-                        {
-                            if(!mSneakGamepadShortcut && mActionManager->isSneaking())
-                                mActionManager->toggleSneaking();
-                            if(mSneakToggleShortcutTimer <= 0.3f)
-                                mSneakToggleShortcutTimer += dt;
-                        }
-                    }
-                    else
+                    if(!mControllerManager->joystickLastUsed())
                         player.setSneak(actionIsActive(A_Sneak));
                 }
 
@@ -537,13 +354,6 @@ namespace MWInput
                             MWBase::Environment::get().getWorld()->togglePOV();
                         }
                         mPreviewPOVDelay = 0.f;
-                        mGamepadZoom = 0;
-                    }
-
-                    if(mGamepadZoom)
-                    {
-                        MWBase::Environment::get().getWorld()->changeVanityModeScale(mGamepadZoom);
-                        MWBase::Environment::get().getWorld()->setCameraDistance(mGamepadZoom, true, true);
                     }
                 }
             }
@@ -562,8 +372,6 @@ namespace MWInput
                 updateIdleTime(dt);
             }
         }
-        else
-            mGamepadZoom = 0;
         mAttemptJump = false; // Can only jump on first frame input is on
     }
 
@@ -574,17 +382,18 @@ namespace MWInput
 
     void InputManager::setGamepadGuiCursorEnabled(bool enabled)
     {
-        mGamepadGuiCursorEnabled = enabled;
+        mControllerManager->setGamepadGuiCursorEnabled(enabled);
     }
 
     void InputManager::changeInputMode(bool guiMode)
     {
         mGuiCursorEnabled = guiMode;
-        mMouseManager->setGuiCursorEnabled(guiMode);
-        mMouseManager->setMouseLookEnabled(!guiMode);
+        mControllerManager->setGuiCursorEnabled(mGuiCursorEnabled);
+        mMouseManager->setGuiCursorEnabled(mGuiCursorEnabled);
+        mMouseManager->setMouseLookEnabled(!mGuiCursorEnabled);
         if (guiMode)
             MWBase::Environment::get().getWindowManager()->showCrosshair(false);
-        MWBase::Environment::get().getWindowManager()->setCursorVisible(guiMode && (!mJoystickLastUsed || mGamepadGuiCursorEnabled));
+        MWBase::Environment::get().getWindowManager()->setCursorVisible(guiMode && (!mControllerManager->joystickLastUsed() || mControllerManager->gamepadGuiCursorEnabled()));
         // if not in gui mode, the camera decides whether to show crosshair or not.
     }
 
@@ -597,9 +406,6 @@ namespace MWInput
         {
             if (it->first == "Input" && it->second == "grab cursor")
                 mGrabCursor = Settings::Manager::getBool("grab cursor", "Input");
-
-            if (it->first == "Input" && it->second == "enable controller")
-                mJoystickEnabled = Settings::Manager::getBool("enable controller", "Input");
 
             if (it->first == "Video" && (
                     it->second == "resolution x"
@@ -686,7 +492,7 @@ namespace MWInput
 
         if (!mControlsDisabled && !consumed)
             mInputBinder->keyPressed (arg);
-        mJoystickLastUsed = false;
+        mControllerManager->setJoystickLastUsed(false);
     }
 
     void InputManager::textInput(const SDL_TextInputEvent &arg)
@@ -699,124 +505,12 @@ namespace MWInput
 
     void InputManager::keyReleased(const SDL_KeyboardEvent &arg )
     {
-        mJoystickLastUsed = false;
+        mControllerManager->setJoystickLastUsed(false);
         OIS::KeyCode kc = mInputManager->sdl2OISKeyCode(arg.keysym.sym);
 
         if (!mInputBinder->detectingBindingState())
             setPlayerControlsEnabled(!MyGUI::InputManager::getInstance().injectKeyRelease(MyGUI::KeyCode::Enum(kc)));
         mInputBinder->keyReleased (arg);
-    }
-
-    void InputManager::buttonPressed(int deviceID, const SDL_ControllerButtonEvent &arg )
-    {
-        if (!mJoystickEnabled || mInputBinder->detectingBindingState())
-            return;
-
-        mJoystickLastUsed = true;
-        if (MWBase::Environment::get().getWindowManager()->isGuiMode())
-        {
-            if (gamepadToGuiControl(arg))
-                return;
-            if (mGamepadGuiCursorEnabled)
-            {
-                // Temporary mouse binding until keyboard controls are available:
-                if (arg.button == SDL_CONTROLLER_BUTTON_A) // We'll pretend that A is left click.
-                {
-                    bool mousePressSuccess = mMouseManager->injectMouseButtonPress(SDL_BUTTON_LEFT);
-                    if (MyGUI::InputManager::getInstance().getMouseFocusWidget())
-                    {
-                        MyGUI::Button* b = MyGUI::InputManager::getInstance().getMouseFocusWidget()->castType<MyGUI::Button>(false);
-                        if (b && b->getEnabled())
-                            MWBase::Environment::get().getWindowManager()->playSound("Menu Click");
-                    }
-
-                    setPlayerControlsEnabled(!mousePressSuccess);
-                }
-            }
-        }
-        else
-            setPlayerControlsEnabled(true);
-
-        //esc, to leave initial movie screen
-        OIS::KeyCode kc = mInputManager->sdl2OISKeyCode(SDLK_ESCAPE);
-        setPlayerControlsEnabled(!MyGUI::InputManager::getInstance().injectKeyPress(MyGUI::KeyCode::Enum(kc), 0));
-
-        if (!mControlsDisabled)
-            mInputBinder->buttonPressed(deviceID, arg);
-    }
-
-    void InputManager::buttonReleased(int deviceID, const SDL_ControllerButtonEvent &arg )
-    {
-        if(mInputBinder->detectingBindingState())
-        {
-            mInputBinder->buttonReleased(deviceID, arg);
-            return;
-        }
-        if (!mJoystickEnabled || mControlsDisabled)
-            return;
-
-        mJoystickLastUsed = true;
-        if (MWBase::Environment::get().getWindowManager()->isGuiMode())
-        {
-            if (mGamepadGuiCursorEnabled)
-            {
-                // Temporary mouse binding until keyboard controls are available:
-                if (arg.button == SDL_CONTROLLER_BUTTON_A) // We'll pretend that A is left click.
-                {
-                    bool mousePressSuccess = mMouseManager->injectMouseButtonRelease(SDL_BUTTON_LEFT);
-                    if (mInputBinder->detectingBindingState()) // If the player just triggered binding, don't let button release bind.
-                        return;
-
-                    setPlayerControlsEnabled(!mousePressSuccess);
-                }
-            }
-        }
-        else
-            setPlayerControlsEnabled(true);
-
-        //esc, to leave initial movie screen
-        OIS::KeyCode kc = mInputManager->sdl2OISKeyCode(SDLK_ESCAPE);
-        setPlayerControlsEnabled(!MyGUI::InputManager::getInstance().injectKeyRelease(MyGUI::KeyCode::Enum(kc)));
-
-        mInputBinder->buttonReleased(deviceID, arg);
-    }
-
-    void InputManager::axisMoved(int deviceID, const SDL_ControllerAxisEvent &arg )
-    {
-        if(!mJoystickEnabled || mControlsDisabled)
-            return;
-
-        mJoystickLastUsed = true;
-        if (MWBase::Environment::get().getWindowManager()->isGuiMode())
-        {
-            gamepadToGuiControl(arg);
-        }
-        else
-        {
-            if(mPreviewPOVDelay == 1.f && arg.value) // Preview Mode Gamepad Zooming
-            {
-                if(arg.axis == SDL_CONTROLLER_AXIS_TRIGGERRIGHT)
-                {
-                    mGamepadZoom = arg.value * 0.85f / 1000.f;
-                    return; // Do not propagate event.
-                }
-                else if(arg.axis == SDL_CONTROLLER_AXIS_TRIGGERLEFT)
-                {
-                    mGamepadZoom = -arg.value * 0.85f / 1000.f;
-                    return; // Do not propagate event.
-                }
-            }
-        }
-        mInputBinder->axisMoved(deviceID, arg);
-    }
-
-    void InputManager::controllerAdded(int deviceID, const SDL_ControllerDeviceEvent &arg)
-    {
-        mInputBinder->controllerAdded(deviceID, arg);
-    }
-    void InputManager::controllerRemoved(const SDL_ControllerDeviceEvent &arg)
-    {
-        mInputBinder->controllerRemoved(arg);
     }
 
     void InputManager::windowFocusChange(bool have_focus)
@@ -1397,5 +1091,15 @@ namespace MWInput
     void InputManager::resetToDefaultControllerBindings()
     {
         loadControllerDefaults(true);
+    }
+
+    void InputManager::setJoystickLastUsed(bool enabled)
+    {
+        mControllerManager->setJoystickLastUsed(enabled);
+    }
+
+    bool InputManager::joystickLastUsed()
+    {
+        return mControllerManager->joystickLastUsed();
     }
 }
